Migrate from Gulp + Webpack to Vite.
