+++ package.json
@@ -9,19 +9,15 @@
         "npm": ">=8.5.0"
     },
     "scripts": {
-        "dev": "supervisor -w Gulpfile.js,webpack.config.js,tsconfig.json supervisor -w Gulpfile.js -x gulp --",
+        "dev": "npx tsx server.ts",
         "prettier": "prettier --write \"src/**/*.{ts,tsx}\"",
         "prettier:check": "prettier --check \"src/**/*.{ts,tsx}\"",
-        "webpack": "webpack",
-        "webpack-watch": "webpack --watch --progress --color",
         "lint": "eslint src/",
         "lintold": "eslint src/ --ext=.ts,.tsx -c .eslintrc.js",
         "lint:fix": "eslint --fix src/ --ext=.ts,.tsx -c .eslintrc.js",
         "test": "jest",
         "fresh-test": "clear && jest",
-        "gulp": "gulp",
         "spellcheck": "cspell \"src/**/*.{ts,tsx}\"",
-        "minify-index": "gulp minify-index --silent",
         "analyze": "webpack --mode production --profile --optimization-minimize --devtool=source-map",
         "yarn": "yarn"
     },
@@ -49,6 +45,7 @@
         "css-what": "^6.1.0",
         "cssnano": "^6.0.1",
         "deepl-node": "^1.10.1",
+        "esbuild-plugin-react-virtualized": "^1.0.4",
         "eslint": "^9.7.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-import": "^2.29.1",
@@ -62,14 +59,6 @@
         "fork-ts-checker-webpack-plugin": "^9.0.0",
         "globals": "^15.8.0",
         "goban": "=8.3.61",
-        "gulp": "^5.0.0",
-        "gulp-clean-css": "^4.3.0",
-        "gulp-eslint-new": "^2.2.0",
-        "gulp-livereload": "^4.0.2",
-        "gulp-postcss": "^9.0.0",
-        "gulp-rename": "^2.0.0",
-        "gulp-sourcemaps": "^3.0.0",
-        "gulp-stylus": "^2.7.0",
         "html-minifier": "4.0.0",
         "husky": "^8.0.1",
         "jest": "^29.7.0",
@@ -89,9 +78,8 @@
         "ts-jest": "^29.1.1",
         "ts-loader": "^9.5.1",
         "typescript": "=5.5.4",
-        "webpack": "^5.94.0",
-        "webpack-bundle-analyzer": "^4.10.2",
-        "webpack-cli": "^5.1.4",
+        "vite": "^5.4.3",
+        "vite-express": "^0.19.0",
         "xgettext-js": "^3.0.0",
         "yarn": "^1.22.19"
     },
@@ -115,9 +103,11 @@
         "@types/react-dom": "^18.0.11",
         "@types/react-virtualized": "^9.21.21",
         "@types/sanitize-html": "^1.22.0",
+        "@vitejs/plugin-react": "^4.3.1",
         "array-move": "^2.2.1",
         "d3": "^7.6.1",
         "date-fns": "^3.6.0",
+        "eslint-plugin-react-refresh": "^0.4.11",
         "eventemitter3": "^5.0.0",
         "howler": "^2.2.1",
         "jquery": "^3.6.0",
