+++ vite.config.ts
@@ -1,0 +1,23 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import fixReactVirtualized from "esbuild-plugin-react-virtualized";
+import path from "path";
+
+export default defineConfig({
+    root: ".",
+    build: {
+        outDir: "dist",
+        manifest: "src/manifest.json",
+    },
+    plugins: [react()],
+    resolve: {
+        alias: {
+            "@": path.resolve(__dirname, "src"),
+        },
+    },
+    optimizeDeps: {
+        esbuildOptions: {
+            plugins: [fixReactVirtualized],
+        },
+    },
+});
